+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 ## 2.5.26
 
 ### Added
+
 - Ability to search for square 512x512 and 1024x1024 images for capsules and wide capsules (disabled by default). Resolves [issue 703](https://github.com/SteamGridDB/steam-rom-manager/issues/703).
+
+### Changed
+
+- Simplify "Create Parser" interface by starting with "Title Modifier", "Controllers", "Artwork Providers", and "Local Artwork" sections all minimized.
 
 ## 2.5.25